refactor: solving conflicts
