+++ src/user/dto/index.ts
@@ -1,1 +1,2 @@
 export * from './list-user-by-id.dto';
+export * from '../dto/save-user-dto';
