+++ src/user/infra/repositories/index.ts
@@ -1,2 +1,1 @@
 export * from './user.repository';
-export * from './users-repository';
