+++ src/user/interfaces/index.ts
@@ -1,3 +1,2 @@
-export * from './user.repository';
-export * from './users-repository';
 export * from './sign-in';
+export * from './user.repository';
