+++ src/user/user.module.ts
@@ -4,8 +4,8 @@ import { JwtModule } from '@nestjs/jwt';
 
 import { User } from '@/user/infra/entities';
 import { ListUserByIdUseCase } from '@/user/use-cases';
-import { IUserRepository } from '@/user/interfaces';
 import { UserRepository } from '@/user/infra/repositories';
+import { IUserRepository } from './interfaces/user.repository';
 
 @Module({
   imports: [
